+ pomotimer: restart the countdown if received a USR1 signal
this way we can have a keypress that restarts the counter for us.  we
would do it when starting a new pomodoro "session" -- typically 25
minutes for us -- at the terminal, and it would follow taskstart button.
we wouldn't want to combine those into one key though, because we
sometimes switch the clock on and off or switch it but we might be doing
something personal while still sitting in the chair and we want the 25m
timer to continue counting